+++ srai/loaders/osm_loaders/osm_pbf_loader.py
@@ -6,7 +6,7 @@ This module contains loader capable of loading OpenStreetMap features from `*.os
 
 from collections.abc import Iterable
 from pathlib import Path
-from typing import TYPE_CHECKING, Optional, Union
+from typing import TYPE_CHECKING, Literal, Optional, Union
 
 import geopandas as gpd
 from shapely.geometry.base import BaseGeometry
@@ -48,6 +48,7 @@ class OSMPbfLoader(OSMLoader):
         pbf_file: Optional[Union[str, Path]] = None,
         download_source: "OsmExtractSource" = "any",
         download_directory: Union[str, Path] = "files",
+        verbosity_mode: Literal["silent", "transient", "verbose"] = "transient",
     ) -> None:
         """
         Initialize OSMPbfLoader.
@@ -61,11 +62,16 @@ class OSMPbfLoader(OSMLoader):
                 Defaults to "any".
             download_directory (Union[str, Path], optional): Directory where to save the downloaded
                 `*.osm.pbf` files. Ignored if `pbf_file` is provided. Defaults to "files".
+            verbosity_mode (Literal["silent", "transient", "verbose"], optional): Set progress
+                verbosity mode. Can be one of: silent, transient and verbose. Silent disables
+                output completely. Transient tracks progress, but removes output after finished.
+                Verbose leaves all progress outputs in the stdout. Defaults to "transient".
         """
         import_optional_dependencies(dependency_group="osm", modules=["quackosm"])
         self.pbf_file = pbf_file
         self.download_source = download_source
         self.download_directory = download_directory
+        self.verbosity_mode = verbosity_mode
 
     def load(
         self,
@@ -118,14 +124,14 @@ class OSMPbfLoader(OSMLoader):
         pbf_reader = self._get_pbf_file_reader(area_wgs84, tags)
 
         if self.pbf_file is not None:
-            features_gdf = pbf_reader.get_features_gdf(
+            features_gdf = pbf_reader.convert_pbf_to_geodataframe(
                 file_paths=self.pbf_file,
                 keep_all_tags=keep_all_tags,
                 explode_tags=explode_tags,
                 ignore_cache=ignore_cache,
             )
         else:
-            features_gdf = pbf_reader.get_features_gdf_from_geometry(
+            features_gdf = pbf_reader.convert_geometry_to_geodataframe(
                 keep_all_tags=keep_all_tags, explode_tags=explode_tags, ignore_cache=ignore_cache
             )
 
@@ -181,14 +187,14 @@ class OSMPbfLoader(OSMLoader):
         geoparquet_file_path: Path
 
         if self.pbf_file is not None:
-            geoparquet_file_path = pbf_reader.convert_pbf_to_gpq(
+            geoparquet_file_path = pbf_reader.convert_pbf_to_parquet(
                 pbf_path=self.pbf_file,
                 keep_all_tags=keep_all_tags,
                 explode_tags=explode_tags,
                 ignore_cache=ignore_cache,
             )
         else:
-            geoparquet_file_path = pbf_reader.convert_geometry_filter_to_gpq(
+            geoparquet_file_path = pbf_reader.convert_geometry_to_parquet(
                 keep_all_tags=keep_all_tags, explode_tags=explode_tags, ignore_cache=ignore_cache
             )
 
@@ -202,8 +208,9 @@ class OSMPbfLoader(OSMLoader):
 
         pbf_reader = PbfFileReader(
             tags_filter=tags,
-            geometry_filter=area_wgs84.unary_union,
+            geometry_filter=area_wgs84.union_all(),
             working_directory=self.download_directory,
             osm_extract_source=OsmExtractSource(self.download_source),
+            verbosity_mode=self.verbosity_mode,
         )
         return pbf_reader