+++ tests/embedders/hex2vec/constants.py
@@ -1,5 +1,7 @@
 """Constants for hex2vec tests."""
 
+from srai.loaders.osm_loaders.filters.hex2vec import HEX2VEC_FILTER
+
 ENCODER_SIZES = [32, 16]
 TRAINER_KWARGS = {"max_epochs": 1, "accelerator": "cpu", "deterministic": True}
 
@@ -11,6 +13,13 @@ PREDEFINED_TEST_CASES = [
         "h3_res": 9,
         "radius": 7,
         "seed": 42,
+        "tags": {
+            t: HEX2VEC_FILTER[t]
+            for t in [
+                "building",
+                "amenity",
+            ]
+        },
     },
     {
         "test_case_name": "poz_8",
@@ -19,5 +28,12 @@ PREDEFINED_TEST_CASES = [
         "h3_res": 8,
         "radius": 6,
         "seed": 5555,
+        "tags": {
+            t: HEX2VEC_FILTER[t]
+            for t in [
+                "building",
+                "amenity",
+            ]
+        },
     },
 ]